+++ FE/src/App.js
@@ -15,7 +15,6 @@ import { BrowserRouter as Router, Routes, Route, Navigate } from "react-router-d
 // import CreateProject from "./components/CreateProject";
 const CreateProject = React.lazy(() => import("./components/CreateProject"));
 import styled from "styled-components";
-// import TextEditor from "./components/shareMemo/test";
 
 import "./App.css";
 import "./reset.css";
