+++ FE/src/components/recommend/InfiniteScroll.js
@@ -1,6 +1,6 @@
 import { useRef, useEffect, useState } from "react";
-import { Link } from "react-router-dom";
 import styled from "styled-components";
+import ScrollRow from "./ScrollRow";
 
 function Observer({ onIntersect, stopObserver }) {
   // console.log(stopObserver);
@@ -78,43 +78,11 @@ function ScrollView() {
     fetchUploadProjectInfo();
   };
 
-  const ScrollRow = ({ el }) => {
-    const defaultHashTags = ["제주도", "여행"];
-    return (
-      <StyledDiv>
-        <Link to={`project/${el._id}`}>
-          <RecommendItems>
-            <div className="background-img" style={{ backgroundImage: `url(${el.img})` }}>
-              <div className="uploadProjectInfo-title">{el.project_title}</div>
-              <HashTagBackground>
-                <div className="uploadProjectInfo-hashTags">
-                  {el.hashTags.length === 0
-                    ? defaultHashTags.map((hashTag, index) => (
-                        <span key={index}>
-                          {"#"}
-                          {hashTag}{" "}
-                        </span>
-                      ))
-                    : el.hashTags.map((hashTag, index) => (
-                        <span key={index}>
-                          {"#"}
-                          {hashTag}{" "}
-                        </span>
-                      ))}
-                </div>
-              </HashTagBackground>
-            </div>
-          </RecommendItems>
-        </Link>
-      </StyledDiv>
-    );
-  };
-
   return (
     <RecommendContents>
       <ScrollViewContents>
         {uploadProjectInfo.map((el, i) => {
-          return <ScrollRow key={i} el={el} />;
+          return <ScrollRow el={el} key={i} />;
         })}
       </ScrollViewContents>
       <Observer onIntersect={loadMore} stopObserver={stopObserver} />
@@ -122,50 +90,10 @@ function ScrollView() {
   );
 }
 
-const StyledDiv = styled.div`
-  bottom: 0px;
-  transition: bottom 0.3s;
-  margin: 10px 20px 10px 0px;
-
-  &:hover {
-    position: relative;
-    bottom: 10px;
-  }
-`;
-
 const ScrollViewContents = styled.div`
   display: flex;
 `;
 
-const RecommendItems = styled.div`
-  height: 200px;
-  width: 200px;
-  border-radius: 10px;
-  background-color: white;
-  cursor: pointer;
-  background: #232a3c;
-
-  .background-img {
-    height: 200px;
-    width: 200px;
-    border-radius: 10px;
-    box-shadow: 4px 4px 4px rgb(0 0 0 / 25%);
-    background-position: center;
-  }
-
-  .uploadProjectInfo-title {
-    font-size: 25px;
-    color: #f8f9fa;
-    text-align: center;
-    padding-top: 60px;
-  }
-  .uploadProjectInfo-hashTags {
-    font-size: 15px;
-    color: #7c8289;
-    text-align: center;
-    padding: 23.5px;
-  }
-`;
 const HashtagResult = styled.div`
   font-size: 30px;
   font-weight: bold;
@@ -174,13 +102,6 @@ const HashtagResult = styled.div`
 `;
 const HashtagResultText = styled(HashtagResult)`
   color: #f8f9fa;
-`;
-
-const HashTagBackground = styled.div`
-  margin-top: 53px;
-  background: white;
-  border-radius: 0px 0px 10px 10px;
-  height: 62px;
 `;
 
 const RecommendContents = styled.div`
@@ -192,6 +113,11 @@ const RecommendContents = styled.div`
     /* width: 0px;
     height: 7px; */
     display: none;
+  }
+  ::-webkit-scrollbar-thumb {
+    background-color: #232a3c;
+    border-radius: 10px;
+    border: 2px solid transparent;
   }
 `;
